controles and globals separated, uartMidi fix, defaulting caches when preset or mod changes
