+++ controls.h
@@ -1,0 +1,7 @@
+#ifndef CONTROLS_H
+#define CONTROLS_H
+
+void ProcessButtons();
+void ProcessEncoders();
+
+#endif
